+++ app/Models/BillOfMaterial.php
@@ -13,6 +13,8 @@ class BillOfMaterial extends Model
         'measurement_unit',
         'total_cost',
         'active',
+        'created_at',
+        'updated_at',
     ];
 
     public function __construct(array $attributes = [])
@@ -41,4 +43,22 @@ class BillOfMaterial extends Model
     {
         return self::orderBy('created_at', 'asc')->paginate(10);
     }
+
+    public static function SearchOfBillMaterial($keywords = null)
+    {
+        $query = self::query();
+
+        if ($keywords) {
+            $query->where('bom_id', 'LIKE', "%{$keywords}%")
+                  ->orWhere('bom_name', 'LIKE', "%{$keywords}%")
+                  ->orWhere('measurement_unit', 'LIKE', "%{$keywords}%")
+                  ->orWhere('total_cost', 'LIKE', "%{$keywords}%")
+                  ->orWhere('active', 'LIKE', "%{$keywords}%")
+                  ->orWhere('created_at', 'LIKE', "%{$keywords}%")
+                  ->orWhere('updated_at', 'LIKE', "%{$keywords}%");
+        }
+
+        return $query->orderBy('created_at', 'asc')->paginate(10);
+    }
+
 }
